Implement routing and component structure for secret management application
- Introduced a new App component to manage routing and display features for creating and uploading secrets.
- Added CreateSecret and Upload components for handling secret creation and file uploads, respectively.
- Implemented a Prefetcher component to fetch and display secrets based on unique identifiers.
- Enhanced the ConfigProvider to manage application configuration and feature toggles.
- Established a shared context for configuration management and improved the overall application structure.
