+++ next/src/app/providers/ConfigProvider.tsx
@@ -1,0 +1,1 @@
+export { ConfigProvider } from "@shared/context/ConfigContext";
