+++ next/src/features/display-secret/Result.tsx
