+++ next/src/main.tsx
@@ -1,8 +1,8 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
-import "./index.css";
-import App from "./App.tsx";
-import { ConfigProvider } from "./utils/ConfigContext";
+import "@shared/styles/index.css";
+import App from "@app/App.tsx";
+import { ConfigProvider } from "@shared/context/ConfigContext";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
