+++ next/src/shared/lib/random.ts
@@ -1,0 +1,21 @@
+export const randomInt = (min: number, max: number): number => {
+  const byteArray = new Uint8Array(1);
+  window.crypto.getRandomValues(byteArray);
+
+  const range = max - min;
+  const maxRange = 256;
+  if (byteArray[0] >= Math.floor(maxRange / range) * range) {
+    return randomInt(min, max);
+  }
+  return min + (byteArray[0] % range);
+};
+
+export const randomString = (): string => {
+  let text = "";
+  const possible =
+    "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
+  for (let i = 0; i < 22; i++) {
+    text += possible.charAt(randomInt(0, possible.length));
+  }
+  return text;
+};
