+++ next/src/shared/styles/index.css
@@ -3,3 +3,4 @@
   themes: emerald --default, dim;
 
 }
+
